Move to `operator` package
