+++ pkg/streamingpromql/operator/limiting_pool_test.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package pooling
+package operator
 
 import (
 	"testing"
